scripts: west_commands: fix board filter loop in do_run_common()
A loop intended to remove entries where the board name did not match a
board regex did not work as intended because it was deleting entries
from a list while iterating over it. As a result, entries that didn't
match the board ended up being included anyway.

Fix this by iterating over the lists in reverse order, so that the
iteration index is not invalidated by deleting the current element.

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>